Add commit hashes using LanguageModelSession integration
