+++ GitClient/Models/SearchArguments.swift
@@ -1,0 +1,18 @@
+//
+//  SearchArguments.swift
+//  GitClient
+//
+//  Created by Makoto Aoyama on 2025/06/23.
+//
+
+import Foundation
+
+struct SearchArguments {
+    var revisionRange = ""
+    var grep: [String] = []
+    var grepAllMatch = false
+    var s = ""
+    var g = ""
+    var authors:[String] = []
+    var paths: [String] = []
+}
